ResponseStatusException First Commit

+++ src/main/java/com/stacksimplify/restservices/services/UserService.java
@@ -1,9 +1,13 @@
 package com.stacksimplify.restservices.services;
 
 import com.stacksimplify.restservices.entities.User;
+import com.stacksimplify.restservices.exceptions.UserExistsException;
+import com.stacksimplify.restservices.exceptions.UserNotFoundException;
 import com.stacksimplify.restservices.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,22 +24,41 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User createUser(User user) {
+    public User createUser(User user) throws UserExistsException {
+        //if user exist using username
+        User existingUser = userRepository.findByUsername(user.getUsername());
+
+        //if not exists throw UserExistsException
+        if(existingUser != null) {
+            throw new UserExistsException("User already exists in repository");
+        }
         return userRepository.save(user);
     }
 
-    public Optional<User> getUserById(Long id) {
+    public Optional<User> getUserById(Long id) throws UserNotFoundException {
         Optional<User> user = userRepository.findById(id);
+        if(!user.isPresent()) {
+            throw new UserNotFoundException("User Not found in user Repository");
+        }
         return user;
     }
 
     //updateUserById
-    public User updateUserById(Long id, User user) {
+    public User updateUserById(Long id, User user) throws UserNotFoundException {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if(!optionalUser.isPresent()) {
+            throw new UserNotFoundException("User Not found in user Repository, provide the correct user id");
+        }
+
         user.setId(id);
         return userRepository.save(user);
     }
 
     public void deleteUserById(Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if(!optionalUser.isPresent()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"User Not found in user Repository, provide the correct user id");
+        }
         if(userRepository.findById(id).isPresent()) {
             userRepository.deleteById(id);
         }
